Prorate Delete Fix
Type(s): FIX

Issue(s):
- JIRA: B2BBE-1053

+++ app/Http/Controllers/B2b/ProrateController.php
@@ -186,11 +186,14 @@ class ProrateController extends Controller
         /** @var Member $manager_member */
         $manager_member = $request->manager_member;
         $this->setModifier($manager_member);
+        $not_found_counter = 0;
         foreach ($request->business_member_leave_type_ids as $id) {
             /**@var BusinessMemberLeaveType $business_member_leave_type */
             $business_member_leave_type = $this->businessMemberLeaveTypeRepo->find($id);
+            if (!$business_member_leave_type) {$not_found_counter++; continue;}
             $this->businessMemberLeaveTypeRepo->delete($business_member_leave_type);
         }
-        return api_response($request, null, 200);
+        $message = $not_found_counter > 0 ? 'Some of prorates have not found': null;
+        return api_response($request, null, 200, ['message' => $message]);
     }
 }
